Add initial version of the lift bottom
This is an experimental two part device, with one part being static
& providing rail in which the other part slides.

The distance between the parts is controlled by a 30 M4 bolt & nut.

The idea is to have 2 triangular surfaces moving against each other,
so that an upper wedge can be moved up and down by rotating
the horizontal bolt.